Add Firebase Hosting deployment workflows and configuration
This commit adds two new files: firebase-hosting-merge.yml and firebase-hosting-pull-request.yml. These files contain the necessary configurations for deploying the project to Firebase Hosting on merge and pull request events. The workflows use the FirebaseExtended/action-hosting-deploy action to deploy the project to Firebase Hosting. The firebase.json file is also updated to include the hosting configuration, specifying the "build" directory as the public folder and setting up a rewrite rule to redirect all requests to the index.html file.